+++ cypress/integration/examples/test2.spec.js
@@ -1,7 +1,0 @@
-/// <reference types="cypress" />
-
-context('test2', () => {
-  it('waits for 2s', () => {
-    cy.wait(2000);
-  });
-});