+++ cypress/integration/examples/test3.spec.js
@@ -1,7 +1,0 @@
-/// <reference types="cypress" />
-
-context('test3', () => {
-  it('waits for 3s', () => {
-    cy.wait(3000);
-  });
-});