+++ cypress/integration/examples/test4.spec.js
@@ -1,7 +1,0 @@
-/// <reference types="cypress" />
-
-context('test4', () => {
-  it('waits for 4s', () => {
-    cy.wait(4000);
-  });
-});