+++ cypress/integration/examples/test5.spec.js
@@ -1,7 +1,0 @@
-/// <reference types="cypress" />
-
-context('test5', () => {
-  it('waits for 5s', () => {
-    cy.wait(5000);
-  });
-});